feat: syn scan mode

+++ modules/scan/connect/connect.go
@@ -6,7 +6,8 @@ import (
 	"time"
 )
 
-func Connect(ip string, port int) (string, int, error) {
+// Connect connect scan
+func Connect(ip string, port uint16) (string, uint16, error) {
 	conn, err := net.DialTimeout("tcp", fmt.Sprintf("%v:%v", ip, port), 2*time.Second)
 
 	defer func() {
